Build: Migrate sources to ES modules, use a Rollup-based build system
The new Rollup-based build system is a very simplified version of the jQuery
one, without all the complex custom compilation aspects.

Also:
* add grunt-compare-size
* deduplicate camelCase implementations
