+++ .eslintignore
@@ -2,7 +2,5 @@ coverage
 external
 node_modules
 *.min.js
-src/intro.js
-src/outro.js
 test/data/jquery-*.js
 test/data/jquery.mobile-*.js
