+++ .eslintrc-browser.json
@@ -18,13 +18,11 @@
 	"env": {},
 
 	"globals": {
-		"window": true,
-		"define": true,
-		"module": true
+		"window": true
 	},
 
 	"rules": {
-		"one-var": ["error", {"var": "always"}],
-		"strict": ["error", "function"]
+		"one-var": [ "error", { "var": "always" } ],
+		"strict": [ "error", "function" ]
 	}
 }
