+++ .eslintrc-node.json
@@ -4,7 +4,7 @@
 	"extends": "jquery",
 
 	"parserOptions": {
-		"ecmaVersion": 2017
+		"ecmaVersion": 2018
 	},
 
 	"env": {
