+++ .gitignore
@@ -7,6 +7,7 @@ CDN
 *.patch
 /*.html
 .DS_Store
+.sizecache.json
 
 # Ignore everything in dist folder except for eslint config
 /dist/*
