+++ package.json
@@ -28,14 +28,17 @@
 	"devDependencies": {
 		"chalk": "2.4.2",
 		"commitplease": "3.2.0",
-		"eslint-config-jquery": "2.0.0",
+		"eslint": "6.8.0",
+		"eslint-config-jquery": "3.0.0",
+		"eslint-plugin-import": "2.20.2",
 		"grunt": "1.0.4",
 		"grunt-cli": "1.3.2",
-		"grunt-contrib-concat": "1.0.1",
+		"grunt-compare-size": "0.4.2",
 		"grunt-contrib-uglify": "4.0.1",
 		"grunt-contrib-watch": "1.1.0",
 		"grunt-eslint": "22.0.0",
 		"grunt-karma": "3.0.2",
+		"gzip-js": "0.3.2",
 		"jquery": "3.5.0",
 		"karma": "4.1.0",
 		"karma-browserstack-launcher": "1.5.1",
@@ -44,6 +47,7 @@
 		"karma-qunit": "3.1.2",
 		"load-grunt-tasks": "5.0.0",
 		"qunit": "2.9.2",
+		"rollup": "2.7.6",
 		"testswarm": "1.1.0"
 	},
 	"keywords": [
