+++ src/.eslintrc.json
@@ -3,29 +3,49 @@
 
 	"extends": "../.eslintrc-browser.json",
 
+	"parserOptions": {
+		"ecmaVersion": 2015,
+		"sourceType": "module"
+	},
+
+
+	"plugins": [ "import" ],
+
 	"rules": {
-		// Strict mode is only enabled in the built file.
-		"strict": "off"
+		"import/extensions": [ "error", "always" ],
+		"import/no-cycle": "error"
 	},
 
 	"globals": {
-		"Proxy": true,
-		"Reflect": true,
-
-		"jQuery": true,
-		"migrateWarn": true,
-		"migrateWarnFunc": true,
-		"migrateWarnProp": true,
-		"jQueryVersionSince": true,
-
-		"noGlobal": true
+		"jQuery": false
 	},
 
 	"overrides": [
 		{
 			"files": "wrapper.js",
+			"parserOptions": {
+				"ecmaVersion": 5,
+				"sourceType": "script"
+			},
+			"rules": {
+				"no-unused-vars": "off"
+			},
 			"globals": {
-				"jQuery": false
+				"jQuery": false,
+				"module": true,
+				"define": false
+			}
+		},
+
+		{
+			"files": "css.js",
+			"globals": {
+
+				// Defining Proxy & Reflect globally would create a danger
+				// of using it unguarded in another place, it seems safer
+				// to define it only for this module.
+				"Proxy": false,
+				"Reflect": false
 			}
 		}
 	]
