+++ src/jquery/deferred.js
@@ -1,3 +1,4 @@
+import { migrateWarn } from "../main.js";
 
 var oldDeferred = jQuery.Deferred,
 	tuples = [
