+++ src/jquery/manipulation.js
@@ -1,3 +1,5 @@
+import { migrateWarn } from "../main.js";
+
 var rxhtmlTag = /<(?!area|br|col|embed|hr|img|input|link|meta|param)(([a-z][^\/\0>\x20\t\r\n\f]*)[^>]*)\/>/gi,
 	origHtmlPrefilter = jQuery.htmlPrefilter,
 	makeMarkup = function( html ) {
