--- src/outro.js
@@ -1,2 +1,0 @@
-return jQuery;
-} );
