+++ src/utils.js
@@ -1,0 +1,5 @@
+export function camelCase( string ) {
+	return string.replace( /-([a-z])/g, function( _, letter ) {
+		return letter.toUpperCase();
+	} );
+}
