+++ test/index.html
@@ -23,7 +23,7 @@
 	</script>
 
 	<!-- Version comparisons -->
-	<script src="../src/compareVersions.js"></script>
+	<script src="data/compareVersions.js"></script>
 
 	<!-- Unit test files -->
 	<script src="migrate.js"></script>
